consolidate app.ts/main.ts
